requirements.txt
httpx
Pillow
google-generativeai
langchain-google-genai
streamlit